Issue #332 Code review. Move normalize_log_level to openeo/api/logs.py
