+++ openeo/api/logs.py
@@ -1,3 +1,7 @@
+import logging
+from typing import Optional, Union
+
+
 class LogEntry(dict):
     """
     Log message and info for jobs and services
@@ -46,3 +50,49 @@ class LogEntry(dict):
         return self["level"]
 
     # TODO: add properties for "code", "time", "path", "links" and "data" with sensible defaults?
+
+
+def normalize_log_level(log_level: Optional[Union[int, str]]) -> int:
+    """Convert a log level that may come from user input, so it could be str, int or None, to the integer
+    constants defined in logging.
+
+    :param log_level: the input to be converted.
+    :raises TypeError: when log_level it is neither a str, an int or None.
+    :return: one of the following log level constants from the standard module ``logging``:
+        logging.ERROR, logging.WARNING, logging.INFO, or logging.DEBUG
+    """
+    if log_level is None:
+        return logging.ERROR
+
+    if isinstance(log_level, str):
+        return string_to_log_level(log_level)
+
+    # Now is should be an int
+    if not isinstance(log_level, int):
+        raise TypeError(
+            f"Value for log_level is not an int or str: type={type(log_level)}, value={log_level!r}"
+        )
+
+    return log_level
+
+
+def string_to_log_level(log_level: str) -> int:
+    """Simpler conversion when you know log_level is always a string."""
+    if not isinstance(log_level, str):
+        raise TypeError(
+            f"Parameter 'log_level' must be type str, but it is type {type(log_level)}. Value: log_level={log_level!r}"
+        )
+    if log_level == "":
+        return logging.ERROR
+
+    log_level = log_level.upper()
+    if log_level in ["CRITICAL", "ERROR"]:
+        return logging.ERROR
+    elif log_level == "WARNING":
+        return logging.WARNING
+    elif log_level == "INFO":
+        return logging.INFO
+    elif log_level == "DEBUG":
+        return logging.DEBUG
+
+    return logging.ERROR
