+++ tests/api/test_logs.py
@@ -1,4 +1,6 @@
-from openeo.api.logs import LogEntry
+import logging
+
+from openeo.api.logs import LogEntry, normalize_log_level, string_to_log_level
 import pytest
 
 
@@ -31,3 +33,58 @@ def test_log_entry_basic_dict():
 def test_log_entry_legacy():
     log = LogEntry(id="log01", level="info", message="hello")
     assert log.log_id == "log01"
+
+
+@pytest.mark.parametrize(
+    ["log_level_in", "expected_log_level"],
+    [
+        (None, logging.ERROR),
+        ("", logging.ERROR),
+        (logging.ERROR, logging.ERROR),
+        ("error", logging.ERROR),
+        ("ERROR", logging.ERROR),
+        (logging.WARNING, logging.WARNING),
+        ("warning", logging.WARNING),
+        ("WARNING", logging.WARNING),
+        (logging.INFO, logging.INFO),
+        ("INFO", logging.INFO),
+        ("info", logging.INFO),
+        (logging.DEBUG, logging.DEBUG),
+        ("DEBUG", logging.DEBUG),
+        ("debug", logging.DEBUG),
+    ],
+)
+def test_normalize_log_level(log_level_in, expected_log_level):
+    assert normalize_log_level(log_level_in) == expected_log_level
+
+
+@pytest.mark.parametrize("log_level", [10.0, b"not a string"])
+def test_normalize_log_level_raises_type_error(log_level):
+    with pytest.raises(TypeError):
+        assert normalize_log_level(log_level)
+
+
+@pytest.mark.parametrize(
+    ["log_level_in", "expected_log_level"],
+    [
+        ("", logging.ERROR),
+        ("error", logging.ERROR),
+        ("ERROR", logging.ERROR),
+        ("warning", logging.WARNING),
+        ("WARNING", logging.WARNING),
+        ("INFO", logging.INFO),
+        ("info", logging.INFO),
+        ("DEBUG", logging.DEBUG),
+        ("debug", logging.DEBUG),
+    ],
+)
+def test_string_to_log_level(log_level_in, expected_log_level):
+    assert string_to_log_level(log_level_in) == expected_log_level
+
+
+@pytest.mark.parametrize(
+    "log_level", [None, 42, 10.0, logging.WARNING, b"not a string"]
+)
+def test_string_to_log_level_raises_type_error(log_level):
+    with pytest.raises(TypeError):
+        assert string_to_log_level(log_level)
